// pronti-frontend-main/front/src/app/modulos/carga-familiar/form/form-carga-familiar.enum.ts
import {SiNoEnum} from '../../../enums/si-no.enum';

export enum FormCargaFamiliarEnum {
  nombres = 'nombres',
  apellidos = 'apellidos',
  parentesco = 'parentesco',
  tipoDocumento = 'tipoDocumento',
  documentoIdentidad = 'documentoIdentidad',
  fechaNacimiento = 'fechaNacimiento',
  discapacidad = 'discapacidad',
  estudia = 'estudia',
  tipoDiscapacidad = 'tipoDiscapacidad',
  genero = 'genero',
  edad = 'edad',
  aplicaUtilidad = 'aplicaUtilidad'

}